Port in header X-Forwarded-Port has highest priority upgrade go chassis
