+++ protocol/http/reverse_proxy.go
@@ -24,7 +24,6 @@ import (
 	"net/http"
 	"time"
 
-	"fmt"
 	"github.com/go-chassis/go-chassis/client/rest"
 	chassisCommon "github.com/go-chassis/go-chassis/core/common"
 	chassisconfig "github.com/go-chassis/go-chassis/core/config"
@@ -33,16 +32,16 @@ import (
 	"github.com/go-chassis/go-chassis/core/invocation"
 	"github.com/go-chassis/go-chassis/core/lager"
 	"github.com/go-chassis/go-chassis/core/loadbalancer"
-	"github.com/go-chassis/go-chassis/core/util/string"
 	"github.com/go-chassis/go-chassis/pkg/runtime"
+	"github.com/go-chassis/go-chassis/pkg/string"
 	"github.com/go-chassis/go-chassis/pkg/util/tags"
 	"github.com/go-chassis/go-chassis/third_party/forked/afex/hystrix-go/hystrix"
-	"github.com/go-mesh/mesher/cmd"
 	"github.com/go-mesh/mesher/common"
 	"github.com/go-mesh/mesher/pkg/egress"
 	"github.com/go-mesh/mesher/pkg/metrics"
 	"github.com/go-mesh/mesher/protocol"
 	"github.com/go-mesh/mesher/resolver"
+	util "github.com/go-mesh/mesher/util"
 	"github.com/go-mesh/openlogging"
 	"strconv"
 	"strings"
@@ -191,7 +190,7 @@ func RemoteRequestHandler(w http.ResponseWriter, r *http.Request) {
 		lager.Logger.Error("Get chain failed: " + err.Error())
 		return
 	}
-	if err = SetLocalServiceAddress(inv, r.Header.Get("X-Forwarded-Port")); err != nil {
+	if err = util.SetLocalServiceAddress(inv, r.Header.Get("X-Forwarded-Port")); err != nil {
 		handleErrorResponse(inv, w, http.StatusBadGateway,
 			err)
 	}
@@ -212,22 +211,6 @@ func RemoteRequestHandler(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-//SetLocalServiceAddress assign invocation endpoint a local service address
-// it uses config in cmd or env fi
-// if it is empty, then try to use original port from client as local port
-func SetLocalServiceAddress(inv *invocation.Invocation, port string) error {
-	inv.Endpoint = cmd.Configs.PortsMap[inv.Protocol]
-	if inv.Endpoint == "" {
-		if port != "" {
-			inv.Endpoint = "127.0.0.1:" + port
-			return nil
-		} else {
-			return fmt.Errorf("[%s] is not supported, [%s] didn't set env [%s] or cmd parameter --service-ports before mesher start",
-				inv.Protocol, inv.MicroServiceName, common.EnvServicePorts)
-		}
-	}
-	return nil
-}
 func copyChassisResp2HttpResp(w http.ResponseWriter, resp *http.Response) {
 	postProcessResponse(resp)
 	copyHeader(w.Header(), resp.Header)
